requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.2.5
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==9.5.0
Werkzeug==2.2.3
zipp==3.15.0
numpy==1.24.3
opencv-python==4.7.0.72
albumentations==1.3.1
gunicorn==19.9.0